backend/requirements.txt
channels==4.0.0
daphne==4.1.2
django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2-binary==2.9.9
